backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.6,<3.0
pymongo>=4.6,<5.0
python-jose[cryptography]>=3.3,<4.0
email-validator>=2.1,<3.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
pdfplumber>=0.11,<0.12
python-docx>=1.1,<2.0
python-multipart>=0.0.9,<1.0
sentence-transformers>=3.0,<4.0
